fix(t9): 实现退格先回退半提交（预编辑恢复为拼音）再删除输入序列的功能
新增 t9WasCommitUndone() 接口用于检测退格是否撤销了半提交状态，
实现退格先回退半提交（预编辑恢复为拼音）再删除输入序列的功能。

添加四个新的集成测试用例：
- test48_undoPartialCommitRestoresPreedit: 测试半提交回退后预编辑恢复
- test49_secondBackspaceIsPlainDelete: 测试第二次退格为普通删除
- test50_plainBackspaceDoesNotUndoCommit: 测试普通退格不撤销提交
- test51_chainedPartialUndo: 测试连续半提交回退功能

JNI 层面扩展 T9DigitBuffer 类，增加 commits_ 记录栈和相关操作方法，
修改 T9Processor 处理退格逻辑以支持半提交撤销机制。

+++ app/src/main/java/com/kingzcheung/xime/ui/keyboard/KeyboardCallbacks.kt
@@ -37,7 +37,11 @@ data class KeyboardCallbacks(
     val onFloatingKeyboardDrag: ((dx: Float, dy: Float) -> Unit)? = null,
     val onFloatingKeyboardDragEnd: (() -> Unit)? = null,
     val onT9ReplaceFullPinyin: ((String) -> Unit)? = null,
-    val onT9RightCommitUndone: ((Int) -> Unit)? = null,
+    /**
+     * 回退最近一次 T9 半提交：清除累积的半提交文本（及输入框中已上屏的文字）。
+     * 服务层根据 t9PartialCommitTexts 计算需回删的文本长度。
+     */
+    val onT9RightCommitUndone: (() -> Unit)? = null,
     /**
      * 右侧候选词即将被 RIME select 前同步通知 T9 控制器。
      * 返回 true 表示控制器判断输入序列已被该候选词完整消费。
